+++ packages/shared/src/types.ts
@@ -73,29 +73,6 @@ export type FetchEsque = (
   init?: RequestInit | RequestInitEsque,
 ) => Promise<ResponseEsque>;
 
-/** This matches the return type from the infra */
-export interface FileData {
-  id: string;
-  createdAt: string;
-
-  fileKey: string | null;
-  fileName: string;
-  fileSize: number;
-  metadata: string | null;
-  customId: string | null;
-
-  callbackUrl: string;
-  callbackSlug: string;
-}
-
-export type UploadedFile = {
-  name: string;
-  key: string;
-  url: string;
-  size: number;
-  customId: string | null;
-};
-
 type PowOf2 = 1 | 2 | 4 | 8 | 16 | 32 | 64 | 128 | 256 | 512 | 1024;
 export type SizeUnit = "B" | "KB" | "MB" | "GB";
 export type FileSize = `${PowOf2}${SizeUnit}`;
