+++ packages/uploadthing/src/fastify.ts
@@ -17,9 +17,10 @@ import {
 import { incompatibleNodeGuard } from "./internal/incompat-node-guard";
 import { initLogger } from "./internal/logger";
 import { toWebRequest } from "./internal/node-http/toWebRequest";
-import type { FileRouter, RouterWithConfig } from "./internal/types";
+import type { RouteHandlerOptions } from "./internal/types";
 import type { CreateBuilderOptions } from "./internal/upload-builder";
 import { createBuilder } from "./internal/upload-builder";
+import type { FileRouter } from "./types";
 
 export type { FileRouter };
 export { UTFiles } from "./internal/types";
@@ -36,7 +37,7 @@ export const createUploadthing = <TErrorShape extends Json>(
 
 export const createRouteHandler = <TRouter extends FileRouter>(
   fastify: FastifyInstance,
-  opts: RouterWithConfig<TRouter>,
+  opts: RouteHandlerOptions<TRouter>,
   done: (err?: Error) => void,
 ) => {
   initLogger(opts.config?.logLevel);
