+++ packages/uploadthing/src/internal/upload-builder.ts
@@ -1,18 +1,144 @@
 import type {
+  ErrorMessage,
   FileRouterInputConfig,
   Json,
+  MaybePromise,
+  Simplify,
   UploadThingError,
 } from "@uploadthing/shared";
 
-import { defaultErrorFormatter } from "./error-formatter";
 import type {
-  AnyParams,
-  MiddlewareFnArgs,
-  UnsetMarker,
-  UploadBuilder,
-  UploadBuilderDef,
-  Uploader,
-} from "./types";
+  FileUploadData,
+  FileUploadDataWithCustomId,
+  UploadedFileData,
+} from "../types";
+import { defaultErrorFormatter } from "./error-formatter";
+import type { JsonParser } from "./parser";
+import type { MiddlewareFnArgs, UTFiles } from "./types";
+
+const unsetMarker = "unsetMarker" as "unsetMarker" & {
+  __brand: "unsetMarker";
+};
+export type UnsetMarker = typeof unsetMarker;
+
+export type ValidMiddlewareObject = {
+  [UTFiles]?: FileUploadDataWithCustomId[];
+  [key: string]: unknown;
+};
+
+type ResolverOptions<TParams extends AnyParams> = {
+  metadata: Simplify<
+    TParams["_metadata"] extends UnsetMarker
+      ? undefined
+      : Omit<TParams["_metadata"], typeof UTFiles>
+  >;
+
+  file: UploadedFileData;
+};
+
+export interface AnyParams {
+  _input: any;
+  _metadata: any; // imaginary field used to bind metadata return type to an Upload resolver
+  _middlewareArgs: MiddlewareFnArgs<any, any, any>;
+  _errorShape: any;
+  _errorFn: any; // used for onUploadError
+  _output: any;
+}
+
+type MiddlewareFn<
+  TInput extends Json | UnsetMarker,
+  TOutput extends ValidMiddlewareObject,
+  TArgs extends MiddlewareFnArgs<any, any, any>,
+> = (
+  opts: TArgs & {
+    files: FileUploadData[];
+    input: TInput extends UnsetMarker ? undefined : TInput;
+  },
+) => MaybePromise<TOutput>;
+
+type ResolverFn<TOutput extends Json | void, TParams extends AnyParams> = (
+  opts: ResolverOptions<TParams>,
+) => MaybePromise<TOutput>;
+
+type UploadErrorFn = (input: {
+  error: UploadThingError;
+  fileKey: string;
+}) => void;
+
+export interface UploadBuilder<TParams extends AnyParams> {
+  input: <TParser extends JsonParser>(
+    parser: TParams["_input"] extends UnsetMarker
+      ? TParser
+      : ErrorMessage<"input is already set">,
+  ) => UploadBuilder<{
+    _input: TParser["_output"];
+    _metadata: TParams["_metadata"];
+    _middlewareArgs: TParams["_middlewareArgs"];
+    _errorShape: TParams["_errorShape"];
+    _errorFn: TParams["_errorFn"];
+    _output: UnsetMarker;
+  }>;
+  middleware: <TOutput extends ValidMiddlewareObject>(
+    fn: TParams["_metadata"] extends UnsetMarker
+      ? MiddlewareFn<TParams["_input"], TOutput, TParams["_middlewareArgs"]>
+      : ErrorMessage<"middleware is already set">,
+  ) => UploadBuilder<{
+    _input: TParams["_input"];
+    _metadata: TOutput;
+    _middlewareArgs: TParams["_middlewareArgs"];
+    _errorShape: TParams["_errorShape"];
+    _errorFn: TParams["_errorFn"];
+    _output: UnsetMarker;
+  }>;
+  onUploadComplete: <TOutput extends Json | void>(
+    fn: ResolverFn<TOutput, TParams>,
+  ) => Uploader<{
+    _input: TParams["_input"];
+    _metadata: TParams["_metadata"];
+    _middlewareArgs: TParams["_middlewareArgs"];
+    _errorShape: TParams["_errorShape"];
+    _errorFn: TParams["_errorFn"];
+    _output: TOutput;
+  }>;
+  onUploadError: (
+    fn: TParams["_errorFn"] extends UnsetMarker
+      ? UploadErrorFn
+      : ErrorMessage<"onUploadError is already set">,
+  ) => UploadBuilder<{
+    _input: TParams["_input"];
+    _metadata: TParams["_metadata"];
+    _middlewareArgs: TParams["_middlewareArgs"];
+    _errorShape: TParams["_errorShape"];
+    _errorFn: UploadErrorFn;
+    _output: UnsetMarker;
+  }>;
+}
+
+export type UploadBuilderDef<TParams extends AnyParams> = {
+  routerConfig: FileRouterInputConfig;
+  inputParser: JsonParser;
+  // eslint-disable-next-line @typescript-eslint/ban-types
+  middleware: MiddlewareFn<TParams["_input"], {}, TParams["_middlewareArgs"]>;
+  errorFormatter: (err: UploadThingError) => TParams["_errorShape"];
+  onUploadError: UploadErrorFn;
+};
+
+export interface Uploader<TParams extends AnyParams> {
+  _def: TParams & UploadBuilderDef<TParams>;
+  resolver: ResolverFn<TParams["_output"], TParams>;
+}
+
+type InOut<
+  TMiddlewareArgs extends MiddlewareFnArgs<any, any, any>,
+  TErrorShape extends Json = { message: string },
+> = (input: FileRouterInputConfig) => UploadBuilder<{
+  _input: UnsetMarker;
+  _metadata: UnsetMarker;
+  _middlewareArgs: TMiddlewareArgs;
+  _errorShape: TErrorShape;
+  _errorFn: UnsetMarker;
+  _output: UnsetMarker;
+}>;
 
 function internalCreateBuilder<
   TMiddlewareArgs extends MiddlewareFnArgs<any, any, any>,
@@ -77,18 +203,6 @@ function internalCreateBuilder<
     },
   };
 }
-
-type InOut<
-  TMiddlewareArgs extends MiddlewareFnArgs<any, any, any>,
-  TErrorShape extends Json = { message: string },
-> = (input: FileRouterInputConfig) => UploadBuilder<{
-  _input: UnsetMarker;
-  _metadata: UnsetMarker;
-  _middlewareArgs: TMiddlewareArgs;
-  _errorShape: TErrorShape;
-  _errorFn: UnsetMarker;
-  _output: UnsetMarker;
-}>;
 
 export type CreateBuilderOptions<TErrorShape extends Json> = {
   errorFormatter: (err: UploadThingError) => TErrorShape;
