+++ packages/uploadthing/src/next.ts
@@ -2,10 +2,11 @@ import type { NextRequest } from "next/server";
 
 import type { Json } from "@uploadthing/shared";
 
-import type { FileRouter, RouterWithConfig } from "./internal/types";
+import type { RouteHandlerOptions } from "./internal/types";
 import type { CreateBuilderOptions } from "./internal/upload-builder";
 import { createBuilder } from "./internal/upload-builder";
 import { INTERNAL_DO_NOT_USE_createRouteHandlerCore } from "./server";
+import type { FileRouter } from "./types";
 
 export type { FileRouter };
 export { UTFiles } from "./internal/types";
@@ -19,7 +20,7 @@ export const createUploadthing = <TErrorShape extends Json>(
   >(opts);
 
 export const createRouteHandler = <TRouter extends FileRouter>(
-  opts: RouterWithConfig<TRouter>,
+  opts: RouteHandlerOptions<TRouter>,
 ) => {
   const handlers = INTERNAL_DO_NOT_USE_createRouteHandlerCore(
     opts,
